+++ src/lwip_t41.h
@@ -66,6 +66,9 @@ int enet_link_speed();
 // Returns the link duplex mode, true for full and false for half.
 bool enet_link_is_full_duplex();
 
+// Returns whether a crossover cable is detected.
+bool enet_link_is_crossover();
+
 // Outputs a raw ethernet frame. This returns false if frame is NULL or if the
 // length is not in the range 60-(MAX_FRAME_LEN-4) (excludes the FCS (frame
 // check sequence)). This also returns false if Ethernet is not initialized.